refactor(python): split dogfood and assessment responsibilities
