+++ .buildchain/kfd/kfd-2/release-claims.json
@@ -11,7 +11,7 @@
     "version": "4.0.0-alpha.5",
     "channel": "repository-default",
     "tag": "v4.0.0-alpha.5",
-    "sourceSha": "dcf84881b74b0f2375fbe0271861cb468177ffa0"
+    "sourceSha": "4d9903dad830c02b7141a2e679ed7010eb1f138b"
   },
   "claims": [
     {
