+++ developer/sdk/kfd/kfd-1/contract-world.witness.json
@@ -221,9 +221,9 @@
     {
       "name": "kungfu-primitive-catalog",
       "sourcePath": "framework/primitive/kungfu-primitive-catalog.contract.json",
-      "sourceSha256": "a7a10d4b6bf9cef652d5e3bedd094ab68a7572cc6e3e66d6a53812c4c7894a3b",
+      "sourceSha256": "d3e7247976dc70d0ae864f6d70488089c630d682adf025ee03e76a2d184d2b55",
       "artifactPath": "config/primitive/kungfu-primitive-catalog.contract.json",
-      "expectedSha256": "a7a10d4b6bf9cef652d5e3bedd094ab68a7572cc6e3e66d6a53812c4c7894a3b",
+      "expectedSha256": "d3e7247976dc70d0ae864f6d70488089c630d682adf025ee03e76a2d184d2b55",
       "byteForByte": true
     }
   ],
